+++ permissions/sharedMailboxes (new separate entitlements)/importSubPermissions.ps1
@@ -25,9 +25,10 @@ $filterValue = 'HelloID Dynamic Shared Mailbox'
 
 # PowerShell commands to import
 $commands = @(
-    'Get-Mailbox'
-    , 'Get-EXOMailboxPermission'
-    , 'Get-EXORecipientPermission'
+    "Get-User",
+    "Get-EXOMailbox",
+    "Get-EXOMailboxPermission",
+    "Get-EXORecipientPermission"
 )
 
 #region functions
@@ -188,36 +189,48 @@ try {
         Write-Information 'Connected to Microsoft Exchange Online'
     }
 
-    $actionMessage = 'getting all mailboxes from Microsoft Exchange Online'
-    
-    $getAllMailboxesParams = @{
-        ResultSize  = 'Unlimited'
-        ErrorAction = 'Stop'
+    # Docs: https://learn.microsoft.com/en-us/powershell/module/exchangepowershell/get-exomailbox?view=exchange-ps
+    $actionMessage = "getting user mailboxes from Microsoft Exchange Online"
+    $getUserMailboxesParams = @{
+        RecipientTypeDetails = 'UserMailbox'
+        ResultSize           = 'Unlimited'
+        Properties           = 'GrantSendOnBehalfTo'
+        ErrorAction          = 'Stop'
     }
-    
-    $mailboxes = Get-Mailbox @getAllMailboxesParams
-    $userMailboxes = $mailboxes | Where-Object { $_.RecipientTypeDetails -eq 'UserMailbox' } | Select-Object Guid, Name, UserPrincipalName, ExternalDirectoryObjectId, GrantSendOnBehalfTo
+
+    $userMailboxes = Get-EXOMailbox @getUserMailboxesParams | Select-Object Guid, Name, UserPrincipalName, ExternalDirectoryObjectId, GrantSendOnBehalfTo
+
     $userMailboxesUpnGrouped = $userMailboxes | Group-Object -Property 'UserPrincipalName' -AsHashTable -AsString
     $userMailboxesGuidGrouped = $userMailboxes | Group-Object -Property 'Guid' -AsHashTable -AsString
     $userMailboxesNameGrouped = $userMailboxes | Group-Object -Property 'Name' -AsHashTable -AsString
     Write-Information "Successfully queried [$($userMailboxes.count)] user mailboxes"
-    $sharedMailboxes = $mailboxes | Where-Object { $_.RecipientTypeDetails -eq 'SharedMailbox' } | Select-Object DisplayName, Name, Guid, UserPrincipalName, GrantSendOnBehalfTo
+
+    # Cleanup for memory
+    $userMailboxes = $null
+
+    # Docs: https://learn.microsoft.com/en-us/powershell/module/exchangepowershell/get-exomailbox?view=exchange-ps
+    $actionMessage = "getting shared mailboxes from Microsoft Exchange Online"
+    $getSharedMailboxesParams = @{
+        RecipientTypeDetails = 'SharedMailbox'
+        ResultSize           = 'Unlimited'
+        Properties           = 'GrantSendOnBehalfTo'
+        ErrorAction          = 'Stop'
+    }
+
+    $sharedMailboxes = Get-EXOMailbox @getSharedMailboxesParams | Select-Object DisplayName, Name, Guid, UserPrincipalName, GrantSendOnBehalfTo
+
+    Write-Information "Successfully queried [$($sharedMailboxes.count)] shared mailboxes"
     
     # Filter shared mailboxes with the matching filter criteria
     if ($filterField -eq 'DisplayName') {
         $sharedMailboxes = $sharedMailboxes | Where-Object { $_.DisplayName -like "$filterValue*" }
     }
     else {
-        # Get mailbox details including custom attributes
-        $allSharedMailboxesDetailed = $mailboxes | Where-Object { $_.RecipientTypeDetails -eq 'SharedMailbox' } | Select-Object DisplayName, Name, Guid, UserPrincipalName, GrantSendOnBehalfTo, CustomAttribute2
-        $sharedMailboxes = $allSharedMailboxesDetailed | Where-Object { $_.$filterField -eq $filterValue } | Select-Object DisplayName, Name, Guid, UserPrincipalName, GrantSendOnBehalfTo
+        $sharedMailboxes = $sharedMailboxes | Where-Object { $_.$filterField -eq $filterValue }
     }
     
     Write-Information "Successfully queried [$($sharedMailboxes.count)] shared mailboxes matching filter [$filterField = $filterValue]"
-    # Cleanup for memory
-    $userMailboxes = $null
-    $mailboxes = $null
-
+  
     # Query all Send As permissions once
     $actionMessage = "getting all recipient permissions from Microsoft Exchange Online"
     $getAllRecipientPermissionsParams = @{
@@ -261,6 +274,8 @@ try {
         }
         
         $numberOfAccounts = $fullAccessUsers.Count
+        $numberOfFullAccess += $numberOfAccounts
+
         $permission = @{
             PermissionReference      = @{
                 Reference = $fullAccessMetadata.Reference
@@ -296,6 +311,8 @@ try {
         }
         
         $numberOfAccounts = $sendAsUsers.Count
+        $numberOfSendAs += $numberOfAccounts
+
         $permission = @{
             PermissionReference      = @{
                 Reference = $sendAsMetadata.Reference
@@ -352,6 +369,8 @@ try {
         }
         
         $numberOfAccounts = $sendOnBehalfUsers.Count
+        $numberOfSendOnBehalf += $numberOfAccounts
+
         $permission = @{
             PermissionReference      = @{
                 Reference = $sendOnBehalfMetadata.Reference
@@ -376,7 +395,7 @@ try {
         #endregion Send On Behalf Permission
     }
 
-    Write-Information 'Exchange Online shared mailbox permission entitlement import completed (separate entitlements)'
+    Write-Information "Target permission import for shared mailboxes completed. Full Access: [$numberOfFullAccess] | Send As: [$numberOfSendAs] | Send on Behalf: [$numberOfSendOnBehalf]"
 }
 catch {
     $ex = $PSItem